requirements.txt
pyspark>=3.0.0
scikit-learn>=0.24.0
numpy>=1.19.0
pandas>=1.1.0
joblib>=1.0.0
pyspark>=3.0.0
kafka-python>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
dash>=2.0.0
plotly>=5.0.0
flask>=2.0.0
requests>=2.25.0
schedule>=1.1.0